Crate minject

The minject-rs library

This library provides a simple API for injecting code (in the form of a DLL) into another process on the Windows platform.

Modules

process
This module provides a drop-in replacement for most of the functionality inside the std::process module.

Macros

initializer!
Creates a suitable initialization wrapper function around the given function.

Structs

Module
A description of a module (DLL) to be injected into a process.

ModuleBuilder
A module builder for a module without an initialization function.

ModuleBuilderWithInit
A module builder for a module with an initialization function.

Shared
A helper type for deserializing objects that can be constructed from a raw handle.

Enums

Error
An error that can occur during the injection process.

InitError
An error that can occur in a call to an initializer function.