Enum minject::Error

pub enum Error {
    Bitness,
    LoadFailed(Error),
    InitNotFound(Error),
    InitError(Option<InitError>),
    Deserialize(DeserializeError),
    UnexpectedExitCode(u32),
    Io(Error),
}

An error that can occur during the injection process.

Variants

Bitness

The target process's bitness (32-bit vs 64-bit) does not match the current process's bitness.

LoadFailed

The module could not be loaded into the target process.

InitNotFound

The module's initializer function was not found.

InitError

An error occurred when the initializer function was called.

Deserialize

An error occurred while deserializing the error message.

UnexpectedExitCode

The remote injection thread returned an unexpected exit code and probably crashed.

Io

An I/O error occurred.